requirements.txt
pandas==2.0.3
datasets==3.1.0
transformers==4.46.3
accelerate==1.0.1
scikit-learn==1.3.2
azureml-mlflow==1.56.0
azureml-core==1.55.0
mlflow==2.12.1
joblib==1.3.2